+++ zellij-utils/src/kdl/mod.rs
@@ -2778,6 +2778,9 @@ impl Options {
         let advanced_mouse_actions =
             kdl_property_first_arg_as_bool_or_error!(kdl_options, "advanced_mouse_actions")
                 .map(|(v, _)| v);
+        let mouse_scroll_resize =
+            kdl_property_first_arg_as_bool_or_error!(kdl_options, "mouse_scroll_resize")
+                .map(|(v, _)| v);
         let mouse_hover_effects =
             kdl_property_first_arg_as_bool_or_error!(kdl_options, "mouse_hover_effects")
                 .map(|(v, _)| v);
@@ -2900,6 +2903,7 @@ impl Options {
             show_startup_tips,
             show_release_notes,
             advanced_mouse_actions,
+            mouse_scroll_resize,
             mouse_hover_effects,
             visual_bell,
             focus_follows_mouse,
@@ -4067,6 +4071,31 @@ impl Options {
             None
         }
     }
+    fn mouse_scroll_resize_to_kdl(&self, add_comments: bool) -> Option<KdlNode> {
+        let comment_text = format!(
+            "{}\n{}\n{}",
+            " ", "// Whether Ctrl+ScrollWheel resizes panes", "// default is true",
+        );
+
+        let create_node = |node_value: bool| -> KdlNode {
+            let mut node = KdlNode::new("mouse_scroll_resize");
+            node.push(KdlValue::Bool(node_value));
+            node
+        };
+        if let Some(mouse_scroll_resize) = self.mouse_scroll_resize {
+            let mut node = create_node(mouse_scroll_resize);
+            if add_comments {
+                node.set_leading(format!("{}\n", comment_text));
+            }
+            Some(node)
+        } else if add_comments {
+            let mut node = create_node(false);
+            node.set_leading(format!("{}\n// ", comment_text));
+            Some(node)
+        } else {
+            None
+        }
+    }
     fn mouse_hover_effects_to_kdl(&self, add_comments: bool) -> Option<KdlNode> {
         let comment_text = format!(
             "{}\n{}\n{}",
@@ -4497,6 +4526,9 @@ impl Options {
         }
         if let Some(advanced_mouse_actions) = self.advanced_mouse_actions_to_kdl(add_comments) {
             nodes.push(advanced_mouse_actions);
+        }
+        if let Some(mouse_scroll_resize) = self.mouse_scroll_resize_to_kdl(add_comments) {
+            nodes.push(mouse_scroll_resize);
         }
         if let Some(mouse_hover_effects) = self.mouse_hover_effects_to_kdl(add_comments) {
             nodes.push(mouse_hover_effects);